+++ src/components/Register/Register.jsx
@@ -1,34 +1,10 @@
 import React from 'react'
-import './Register.css'
-import Image from './Sin título-1.png'
 
 const Register = () => {
   return (
-    <>
-    <img src={Image} alt="imagen cabecera del login"></img>
-    <div className='container'>
-      <div className="header">
-        <div className="text"><b>¡Bienvenido de nuevo!</b></div>
-        <p>Por Favor, introduce tus datos para continuar.</p>
-        <div className="underline"></div>
-      </div>
-      <div className="inputs">
-         <div className="input">
-          <input type="email" placeholder="Dirección Email"/>
-        </div>
-        <br></br>
-         <div className="input">
-          <input type="password" placeholder="Password"/>
-        </div>
-      </div>
-      <div className="Forgot-password">¿Olvidaste la contraseña?</div>
+    <div>
+      
     </div>
-    <div className="submit-container">
-      <div className="submit">Entrar</div>
-      <p>¿Nuevo en Applergic?</p>
-      <div className="Login">Crea tu cuenta aquí</div>
-    </div>
-    </>
   )
 }
